requirements/base.txt
Django==1.9.2
django-cities-light==3.2.0
django-crispy-forms==1.6.0
django-rq==0.9.0
dj-database-url
gunicorn
psycopg2
django-configurations
whitenoise
